backend/requirements.txt
django
djangorestframework
django-model-utils #to use extra model classes like  SoftDeletableModel, TimeStampedModel
xlsxwriter